bug fixs (#24)
* bug fix

--Fixed runes of retreat not showing after last update
--Stone of retreat feathermoon spellID was wrong

* changed stones system

* added functions

-added setHeader function
-added getCooldown function
-added dewdropAdd function

* removed a print

* [feature] Added enemy faction SoR toggle

* Updated structure, Added expansion to every SoR

* [bug] fixed menu style not working

* [maintenance] replaced multiple if statements with one for

* [bug] empty menus were visible

* unused variables

removed some unused local variables

* merged addstone function with dewdrop add

* added faction lock to scrolls of retreat

* fix expansion locks

* icon update in the wrong place

* changed it to save favorites per server

* fix portalsdb favorites

* Update portals.lua

* removed unneeded event

* rebuilt the function for class teleports/portals

* changed to local

* added favorites profiles

* adding a new profile now makes it the active profile

* fixed if there was already a favorites table

* fixs/updates

-fixed wrath mage teleports/portals not showing
-updated toc

* fixed merge issue

* recategorized some spells

-added category other teleports/portals spells
-moved teleport moonglade and deathgate to new cat
-added  Atiesh, Greatstaff of the Guardians portal spell to other teleports/portals cat
-moved astral recall to hearthstone category

* fixed mage portals after patch

fixed mage portals after patch

* fixed portals not showing in favorites if there was a teleport on it

* fixed header not showing on hearthstone spells when you dont have a hearthstone item on you

* Update portals.lua

* Revert "Update portals.lua"

This reverts commit a220d8432356adfd497608a7852ec32a4956cc48.

* Revert "fixed header not showing on hearthstone spells when you dont have a hearthstone item on you"

This reverts commit e791ccd5d6526117aa6af3455180c1eda3745942.

* fixed header on hearthstone spells not always showing

* added a button in options to learn all unknown stones

* Update Broker_Portals.toc

* Bug fixes and added zones to stones of retreat

* fixed wrong zone on some stones

* new slash command

- added new slash command /portals macromenu when added to a macro will open the menu where ever the macro is placed on the action bar.
- removed brokerportals header from the menu
- you can now both right or left click to open the menu

* removed duplicate

* Update Broker_Portals.toc

* Update portals.lua

Update with Stones from WoTLK Alpha

* bug fixs

-removed duplicate functions
-fixed profiles not staying selected after reload
-menu now wont close when added or removing from the favorites menu

* rebuilt a lot of the backend

* travel permit now gets deleted after you click where you want to go

* added missing VANITY_SPELL_REFERENCE check any spell type vanity checks

* feature update

- added standalone button
- moved options into an options tab

* fixed unlock alignment

* fixed a bug in my db creator function

* fixed incorrect options setting

* features

- changed how menu position is calculated
- added SettingsCreater

* fixed incorrect variable

* fixed options arrangement

* fixed flight master's whistle

* fixed a bug that stopped the menu loading if you added a mage teleport to favorites

* added raid vanity portals

* raid portals had different spell ids

* added self cast for placeable items/spells

* fixed fel gateway not deleting

* added homebound portal vanity item

* fixed error when there is no item data, and the menus should now show if no stones are learned

* bug fixs

- fixed item cooldown error
- swapped all instances of GetItemInfo to GetItemInfoInstant

* fixed minimap button not showing

* Revert "Merge branch 'main' of https://github.com/anch999/Broker-Portals into broker-main"

This reverts commit 2168a6914f11a4fee6952278978e54265d7b10c3, reversing
changes made to f8bff9e6047c123b3a621ba2cf30fc83a0cd63c4.

* added karazhan stone of retreat

* update version

* added option to hide the stone of retreat zones

* change v to spellID to make it more clear what it is

* turned some repeated if statements into functions

* added 2 nil checks to remove item function

* Update Broker_Portals.toc

* Bug fixs

- Fixed adding items to favorites breaking the menu
- Fixed error after using any items if delete vanity was on

* added discord link to options

* changed frame hider to use setalpha rather then show/hide

* fixed a remove item bug

* auto hide standalone button will now not show if you are in combat

* portals.lua: put repeated code out of if statements

* Update SettingsCreator.lua

* added class check for mage spells on bronzebeard

* added no id message

* Update README.md

---------

+++ Broker_Portals/Utils.lua
@@ -29,17 +29,19 @@ end
 
 -- deletes item from players inventory if value 2 in the items table is set
 function Portals:RemoveItem(arg2)
-	if not self.db.deleteItem then return end
+	if not self.db.deleteItem or not arg2 then return end
         arg2 = self:ConvertCastName(arg2)
         local item = GetItemInfoInstant(self.deleteItem)
-        if strfind(arg2, item:gsub("%-","")) then
+        if item and strfind(arg2, item.name:gsub("%-","")) then
             local found, bag, slot = self:HasItem(self.deleteItem)
-            if found and C_VanityCollection.IsCollectionItemOwned(self.deleteItem) and self:IsRealmbound(bag, slot) then
-                PickupContainerItem(bag, slot)
-                DeleteCursorItem()
+            if found and self:HasVanity(self.deleteItem) and self:IsRealmbound(bag, slot) then
+              ClearCursor()
+              PickupContainerItem(bag, slot)
+              DeleteCursorItem()
             end
             self.deleteItem = nil
         end
+  Timer.After(5, function() self:UnregisterEvent("ZONE_CHANGED_NEW_AREA") end)
 	self:UnregisterEvent("UNIT_SPELLCAST_SUCCEEDED")
 end
 
@@ -109,22 +111,22 @@ local UnknownList = {}
 function Portals:LearnUnknown()
   for i, v in pairs(UnknownList) do
     if not v then return end
-    if not CA_IsSpellKnown(v) and not Portals:HasItem(i) then
+    if not CA_IsSpellKnown(v) and not self:HasItem(i) then
       RequestDeliverVanityCollectionItem(i)
     else
       UnknownList[i] = nil
     end
   end
-  Portals:ScheduleTimer("learnUnknown", .1)
+  self:ScheduleTimer("learnUnknown", .1)
 end
 
 function Portals:LearnUnknownStones()
   for _,v in pairs(VANITY_ITEMS) do
-    if C_VanityCollection.IsCollectionItemOwned(v.itemid) and not CA_IsSpellKnown(v.learnedSpell) and v.name:match("Stone of Retreat") then
+    if self:HasVanity(v.itemid) and not CA_IsSpellKnown(v.learnedSpell) and v.name:match("Stone of Retreat") then
       UnknownList[v.itemid] = v.learnedSpell
     end
   end
-  Portals:LearnUnknown()
+  self:LearnUnknown()
 end
 
 --for a adding a divider to dew drop menus 
@@ -162,21 +164,21 @@ end
 function Portals:OnEnter(button, unlock)
 
     if self.db.autoMenu and not UnitAffectingCombat("player") then
-      Portals:OpenMenu(button, unlock)
+      self:OpenMenu(button, unlock)
     else
       GameTooltip:SetOwner(button, 'ANCHOR_NONE')
-      GameTooltip:SetPoint(Portals:GetTipAnchor(button))
+      GameTooltip:SetPoint(self:GetTipAnchor(button))
       GameTooltip:ClearLines()
 
       GameTooltip:AddLine('Broker Portals')
       GameTooltip:AddDoubleLine(L['RCLICK'], L['SEE_SPELLS'], 0.9, 0.6, 0.2, 0.2, 1, 0.2)
       GameTooltip:AddDoubleLine(L['ALTCLICK'], L['MOVE_SPELLS'], 0.9, 0.6, 0.2, 0.2, 1, 0.2)
       GameTooltip:AddLine(' ')
-      GameTooltip:AddDoubleLine(L['HEARTHSTONE'] .. ': ' .. GetBindLocation(), Portals:GetHearthCooldown(), 0.9, 0.6, 0.2, 0.2, 1,
+      GameTooltip:AddDoubleLine(L['HEARTHSTONE'] .. ': ' .. GetBindLocation(), self:GetHearthCooldown(), 0.9, 0.6, 0.2, 0.2, 1,
         0.2)
 
       if self.db.showItemCooldowns then
-        local cooldowns = Portals:GetItemCooldowns()
+        local cooldowns = self:GetItemCooldowns()
         if cooldowns ~= nil then
           GameTooltip:AddLine(' ')
           for name, cooldown in pairs(cooldowns) do
@@ -191,22 +193,47 @@ end
 
 local worldFrameHook
 function Portals:OpenMenu(button, showUnlock)
-  GameTooltip:Hide()
-  dewdrop:Open(button,
-  'point', function(parent)
-    local point1, _, point2 = self:GetTipAnchor(button)
-    return point1, point2
-  end,
-  'children', function(level, value)
-      Portals:UpdateMenu(level, value, showUnlock)
-  end)
-
-  if not worldFrameHook then
-    WorldFrame:HookScript("OnEnter", function()
-        if dewdrop:IsOpen(button) then
-            dewdrop:Close()
-        end
+    GameTooltip:Hide()
+    dewdrop:Open(button,
+    'point', function(parent)
+      local point1, _, point2 = self:GetTipAnchor(button)
+      return point1, point2
+    end,
+    'children', function(level, value)
+        self:UpdateMenu(level, value, showUnlock)
     end)
-    worldFrameHook = true
+
+    if not worldFrameHook then
+      WorldFrame:HookScript("OnEnter", function()
+          if dewdrop:IsOpen(button) then
+              dewdrop:Close()
+          end
+      end)
+      worldFrameHook = true
+  end
 end
+
+function Portals:CheckFavorites(ID)
+  return not self.favoritesdb[ID] or not self.favoritesdb[ID][1]
+end
+
+function Portals:HasVanityOrSpell(spellID)
+  return CA_IsSpellKnown(spellID) or self:HasVanity(spellID)
+end
+
+function Portals:HasVanity(ID)
+  return C_VanityCollection.IsCollectionItemOwned(VANITY_SPELL_REFERENCE[ID] or ID)
+end
+
+local _, class = UnitClass('player')
+function Portals:IsPortalKnown(spellID)
+  if class == "HERO" then
+    return CA_IsSpellKnown(818045) and CA_IsSpellKnown(spellID)
+  elseif class == "MAGE" then
+    return CA_IsSpellKnown(spellID)
+  end
+end
+
+function Portals:HasVanityOrItem(itemID)
+  return self:HasItem(itemID) or self:HasVanity(itemID)
 end
